+++ packages/nuxt-module/test/fixtures/with-functions/nuxt.config.ts
@@ -1,0 +1,7 @@
+import MyModule from '../../../src/module'
+
+export default defineNuxtConfig({
+  modules: [MyModule],
+  telemetry: { enabled: false },
+  compatibilityDate: '2025-07-11',
+})