+++ packages/dev/CHANGELOG.md
@@ -1,5 +1,81 @@
 # Changelog
 
+## [4.4.4](https://github.com/netlify/primitives/compare/dev-v4.4.3...dev-v4.4.4) (2025-07-21)
+
+
+### Bug Fixes
+
+* handle 304 responses from edge functions ([#364](https://github.com/netlify/primitives/issues/364)) ([08e2543](https://github.com/netlify/primitives/commit/08e254325163707a397c8dddc0761df6aad63fd7))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @netlify/edge-functions bumped from 2.16.0 to 2.16.1
+
+## [4.4.3](https://github.com/netlify/primitives/compare/dev-v4.4.2...dev-v4.4.3) (2025-07-18)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @netlify/edge-functions bumped from 2.15.8 to 2.16.0
+
+## [4.4.2](https://github.com/netlify/primitives/compare/dev-v4.4.1...dev-v4.4.2) (2025-07-17)
+
+
+### Bug Fixes
+
+* **deps:** update netlify packages ([#333](https://github.com/netlify/primitives/issues/333)) ([0fa1f1e](https://github.com/netlify/primitives/commit/0fa1f1e8d717cbc5c797c19b64b4fa4ffbdaf992))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @netlify/blobs bumped from 10.0.5 to 10.0.6
+    * @netlify/dev-utils bumped from 3.3.0 to 4.0.0
+    * @netlify/edge-functions bumped from 2.15.7 to 2.15.8
+    * @netlify/functions bumped from 4.1.11 to 4.1.12
+    * @netlify/headers bumped from 2.0.5 to 2.0.6
+    * @netlify/images bumped from 1.2.1 to 1.2.2
+    * @netlify/redirects bumped from 3.0.5 to 3.0.6
+    * @netlify/runtime bumped from 4.0.9 to 4.0.10
+    * @netlify/static bumped from 3.0.5 to 3.0.6
+
+## [4.4.1](https://github.com/netlify/primitives/compare/dev-v4.4.0...dev-v4.4.1) (2025-07-15)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @netlify/blobs bumped from 10.0.4 to 10.0.5
+    * @netlify/dev-utils bumped from 3.2.2 to 3.3.0
+    * @netlify/edge-functions bumped from 2.15.6 to 2.15.7
+    * @netlify/functions bumped from 4.1.10 to 4.1.11
+    * @netlify/headers bumped from 2.0.4 to 2.0.5
+    * @netlify/images bumped from 1.2.0 to 1.2.1
+    * @netlify/redirects bumped from 3.0.4 to 3.0.5
+    * @netlify/runtime bumped from 4.0.8 to 4.0.9
+    * @netlify/static bumped from 3.0.4 to 3.0.5
+
+## [4.4.0](https://github.com/netlify/primitives/compare/dev-v4.3.8...dev-v4.4.0) (2025-07-15)
+
+
+### Features
+
+* add `images.remoteURLPatterns` option to `NetlifyDev` ([#348](https://github.com/netlify/primitives/issues/348)) ([606e264](https://github.com/netlify/primitives/commit/606e26475c88a47f41929c5548820f2886094b3a))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @netlify/images bumped from 1.1.2 to 1.2.0
+
 ## [4.3.8](https://github.com/netlify/primitives/compare/dev-v4.3.7...dev-v4.3.8) (2025-07-11)