+++ packages/vite-plugin/CHANGELOG.md
@@ -1,5 +1,69 @@
 # Changelog
 
+## [2.4.4](https://github.com/netlify/primitives/compare/vite-plugin-v2.4.3...vite-plugin-v2.4.4) (2025-07-21)
+
+
+### Bug Fixes
+
+* **docs:** update outdated vite plugin readme ([#365](https://github.com/netlify/primitives/issues/365)) ([8ad4e63](https://github.com/netlify/primitives/commit/8ad4e63efd7a6d27c75c8b44a50b4db6c4f22fd7))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @netlify/dev bumped from 4.4.3 to 4.4.4
+
+## [2.4.3](https://github.com/netlify/primitives/compare/vite-plugin-v2.4.2...vite-plugin-v2.4.3) (2025-07-18)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @netlify/dev bumped from 4.4.2 to 4.4.3
+
+## [2.4.2](https://github.com/netlify/primitives/compare/vite-plugin-v2.4.1...vite-plugin-v2.4.2) (2025-07-17)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @netlify/dev bumped from 4.4.1 to 4.4.2
+    * @netlify/dev-utils bumped from ^3.3.0 to ^4.0.0
+
+## [2.4.1](https://github.com/netlify/primitives/compare/vite-plugin-v2.4.0...vite-plugin-v2.4.1) (2025-07-15)
+
+
+### Bug Fixes
+
+* pass `images` option from Vite plugin ([#351](https://github.com/netlify/primitives/issues/351)) ([b4ec7f2](https://github.com/netlify/primitives/commit/b4ec7f2fe674c007f3bbacb49cb97be768e27b0c))
+
+## [2.4.0](https://github.com/netlify/primitives/compare/vite-plugin-v2.3.10...vite-plugin-v2.4.0) (2025-07-15)
+
+
+### Features
+
+* add `@netlify/nuxt` Nuxt module ([#344](https://github.com/netlify/primitives/issues/344)) ([efdf8df](https://github.com/netlify/primitives/commit/efdf8dfb22ffe33b557d2140fd3a3152e23cc95c))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @netlify/dev bumped from 4.4.0 to 4.4.1
+    * @netlify/dev-utils bumped from ^3.2.2 to ^3.3.0
+
+## [2.3.10](https://github.com/netlify/primitives/compare/vite-plugin-v2.3.9...vite-plugin-v2.3.10) (2025-07-15)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @netlify/dev bumped from 4.3.8 to 4.4.0
+
 ## [2.3.9](https://github.com/netlify/primitives/compare/vite-plugin-v2.3.8...vite-plugin-v2.3.9) (2025-07-11)